2019 p14 - refactor and simplify - solve part 1
the old process_one function was way too complicated and long.
I rewrote the top-level loop to use simply named functions,
and then took process_one apart to write the simple functions
one-by-one. I also found the bug that was haunting me, which
was that sometimes i was subtracting some extra and stockpiling it,
but not modifying the current reagent, so too much would get added
back to the stack.